ADEN-12567 Move Phoenix based sites targeting setup to separate file
